requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
cryptography==42.0.2
defusedxml==0.7.1
distlib==0.3.8
dj-rest-auth==5.0.2
Django==5.0
django-allauth==0.61.1
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-js-asset==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
env==0.1.0
environ==1.0
filelock==3.13.1
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
pillow==10.2.0
pipenv==2023.11.15
platformdirs==4.1.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
setuptools==69.0.3
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.25.0